+++ core/src/main/java/org/lemurproject/galago/core/index/source/LengthSource.java
@@ -8,6 +8,6 @@ import org.lemurproject.galago.core.index.stats.FieldStatistics;
  * @author jfoley
  */
 public interface LengthSource extends DiskSource {
-  public long length(int id);
+  public int length(long id);
   public FieldStatistics getStatistics();
 }
